new ALPHA vmf
the alpha means that this is a new stage of development of the map but
it dosent mean that its more playable then the old one. I wil make it
more playable but i try to implement ewerything that i have learned in
making the beta (i also use moast of the beather assets of the beta and
the logic will e copied).